backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
